feat: Add Spaces ElasticSearch Search Connector - MEED-7590 - Meeds-io/meeds#2476
This change will make a cleanup about SpaceService Deprecated APIs in addition to a refactoring to make the SpaceListAccess class as main entry point to request a list of spaces switch a given filter. When the user or the UI requests for space using 'favorite', 'tags' or 'terms', then the 'SpaceListAccess' has to choose the ElasticSearch Storage instead of the Database Storage in order to retrieve the matching Space Identifiers and consequently request the Database (through Cache mechanism) the True Space Data to display to the end user.

+++ component/core/src/main/java/org/exoplatform/social/core/metadata/tag/TagServiceImpl.java
@@ -126,17 +126,18 @@ public class TagServiceImpl implements TagService {
     // Tags to create
     Set<TagName> tagsToCreate = new HashSet<>(tagNames);
     tagsToCreate.removeAll(storedTagNames);
-    createTags(object, tagsToCreate, audienceId, creatorId);
+    if (!tagsToCreate.isEmpty()) {
+      createTags(object, tagsToCreate, audienceId, creatorId);
+    }
 
     // Tags to delete
     Set<TagName> tagsToDelete = new HashSet<>(storedTagNames);
     tagsToDelete.removeAll(tagNames);
-    deleteTags(metadataItems, tagsToDelete);
-    Set<TagName> newTags = tagsToCreate.stream()
-                                       .filter(tagName -> !(storedTagNames.contains(tagName)))
-                                       .collect(Collectors.toSet());
+    if (!tagsToDelete.isEmpty()) {
+      deleteTags(metadataItems, tagsToDelete);
+    }
     try {
-      listenerService.broadcast(TAG_ADDED_EVENT, object, newTags);
+      listenerService.broadcast(TAG_ADDED_EVENT, object, tagsToCreate);
     } catch (Exception e) {
       LOG.error("could not broadcast event", e);
     }
